+++ CondTools/Ecal/src/EcalTPGWeightGroupHandler.cc
@@ -237,7 +237,7 @@ void popcon::EcalTPGWeightGroupHandler::getNewObjects()
 		      
 		      bool set_the_strip=false;
 		      int stripEEId;
-		      for (int istrip=0; istrip<(int)my_StripEcalLogicId_EE.size(); istrip++) {
+		      for (size_t istrip=0; istrip<my_StripEcalLogicId_EE.size(); istrip++) {
 
 			if(!set_the_strip){
